requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
pytesseract==0.3.9
Werkzeug==2.0.3
